Bump version to 1.7.9

+++ TeleSharp/Types/ReplyMarkup/InlineKeyboard/InlineButtonLoginUrl.cs
@@ -9,14 +9,14 @@ namespace TeleSharp.Types.ReplyMarkup
     public class InlineButtonLoginUrl : InlineKeyboardButton
     {
 
-        internal InlineButtonLoginUrl(string text, string url, int id, string forwardText = null) : base(text)
+        internal InlineButtonLoginUrl(string text, string url, long id, string forwardText = null) : base(text)
         {
             Url = url;
             Id = id;
             ForwardText = forwardText;
         }
         public string Url { get; }
-        public int Id { get; }
+        public long Id { get; }
         public string ForwardText { get; }
     }
 }
